refactor(cli): extract modal-prompts/voice mixins from cli.py (shard s4)
